+++ chetah_utils.py
@@ -1,0 +1,28 @@
+import re
+import spacy
+
+# This process references the following two docs and post
+# https://spacy.io/usage/linguistic-features#tokenization
+# https://towardsdatascience.com/setting-up-text-preprocessing-pipeline-using-scikit-learn-and-spacy-e09b9b76758f
+
+# First we need to import the spacy English library
+nlp = spacy.load("en_core_web_sm")
+
+def lemmatize_string(given_string):
+    # then we split into spacy tokens object called a doc
+    doc = nlp(given_string)
+    # next step is removing stop words with spacy's language model
+    # We also want to remove anything that looks like a url, email
+    # Next get rid of new line characters and spaces, and finally punctuation
+    # and anything that is like a phone number
+    # r"(\(?([0-9]+)-?\)?)"
+    pattern = r"(\(?([0-9]+)-?\)?)|[\u200b]|[$+●]|^[a-zA-Z]$|[@/\\<>]|\\{2}|~"
+    lemma_tokens = [x.lemma_.lower() for x in doc if 
+    not x.is_stop
+    and not x.like_url
+    and not x.like_email
+    and not x.is_space
+    and not x.is_punct
+    and not x.like_num
+    and not re.match(pattern,x.text)]
+    return lemma_tokens
