+++ configuration.ini
@@ -1,3 +1,4 @@
 [chetah]
 dataset_path = ./dataset/final_with_cluster.csv
 inv_path = ./dataset/inv_index.json
+doc_path = ./dataset/doc_table.json
